upgraded devDependencies to fix set-value vulnerability
One of the dependencies of the dependencies was using an old version
of the `set-value` package which GitHub/Dependabot flagged as affected
by a security vulnerability:

CVE: https://github.com/advisories/GHSA-4g88-fppr-53pp

Unfortunately debendabot couldn't create a PR automatically because
of other dependencies constraints.

I tried to upgrade only few packages but in the end the only thing
that seemed to work was to upgrade all the devDependencies.

NOTE: For some reason yarn.lock is now losing the integrity SHA.

Part of ticket: https://trello.com/c/PLR5Dft2