forcing the use of minimist >=0.2.1
this is because old versions are affected by this security vulnerability:

https://github.com/advisories/GHSA-vh95-rmgr-6w4m

unfortunatly `optimist`, one of the dependencies of `node-static`,
is old and depend on an old version. There are no newer versions
so I'm forcing the use of a newer version adding `resolutions` to
the `package.json`.

This is not ideal but from what I can understand `minimist` is a
package to parse arguments so hopefully they did't break their
public interface.

I couldn't find any changelog in their repository and the releases
are very light on details:
https://github.com/substack/minimist

Part of https://trello.com/c/PLR5Dft2