downgraded ws to version 7.1 to avoid yarn warnings
after the upgrade of the dependencies yarn was complaing that some
of `ws` peer dependencies were unmet:

```
warning " > ws@7.3.0" has unmet peer dependency "bufferutil@^4.0.1".
warning " > ws@7.3.0" has unmet peer dependency "utf-8-validate@^5.0.2".
```

There are warnings but considering they're related to a WebSockets
package I preferred to err on the side of caution in case they lead
to bugs.

After downgrading to version `~7.1` yarn doesn't complain anymore.

NOTE: Before these dependencies upgrades the project was still using
an older version of `ws` (`7.0.0`) so even after this the net effect is
that the project is using a newer version of `ws`.